# tinder-back/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tinder API",
        "description": "API documentation for the Tinder-like application",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local development server"
        }
    ],
    "paths": {
        "/interests": {
            "get": {
                "tags": [
                    "Interests"
                ],
                "summary": "Get all interests",
                "operationId": "c1c132652a784e8eae7b0cf746767727",
                "responses": {
                    "200": {
                        "description": "List of interests"
                    }
                }
            },
            "post": {
                "tags": [
                    "Interests"
                ],
                "summary": "Create a new interest",
                "operationId": "3d7825139871d3d8d29db6e2a54380f7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Sports"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Interest created successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/interests/{id}": {
            "get": {
                "tags": [
                    "Interests"
                ],
                "summary": "Get a specific interest",
                "operationId": "7525aede6eac965a832459f8eef7e007",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Interest details"
                    },
                    "404": {
                        "description": "Interest not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Interests"
                ],
                "summary": "Update an interest",
                "operationId": "b774cc484e522a94f0f15036b890589e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Music"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Interest updated successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    },
                    "404": {
                        "description": "Interest not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Interests"
                ],
                "summary": "Delete an interest",
                "operationId": "56ddf77e68f4ffa3d6d2198cc5481d36",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Interest deleted successfully"
                    },
                    "404": {
                        "description": "Interest not found"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}